follow insert league

+++ internal/domain/repository/subscription_follow_league_repository.go
@@ -16,15 +16,15 @@ func NewSubscriptionFollowLeagueRepository(db *gorm.DB) *SubscriptionFollowLeagu
 }
 
 type ISubscriptionFollowLeagueRepository interface {
-	Count(int, int) (int64, error)
+	Count(int64, int64) (int64, error)
 	GetAllPaginate(*entity.Pagination) (*entity.Pagination, error)
-	Get(int, int) (*entity.SubscriptionFollowLeague, error)
+	Get(int64, int64) (*entity.SubscriptionFollowLeague, error)
 	Save(*entity.SubscriptionFollowLeague) (*entity.SubscriptionFollowLeague, error)
 	Update(*entity.SubscriptionFollowLeague) (*entity.SubscriptionFollowLeague, error)
 	Delete(*entity.SubscriptionFollowLeague) error
 }
 
-func (r *SubscriptionFollowLeagueRepository) Count(subId, leagueId int) (int64, error) {
+func (r *SubscriptionFollowLeagueRepository) Count(subId, leagueId int64) (int64, error) {
 	var count int64
 	err := r.db.Model(&entity.SubscriptionFollowLeague{}).Where(&entity.SubscriptionFollowLeague{SubscriptionID: int64(subId), LeagueID: int64(leagueId)}).Count(&count).Error
 	if err != nil {
@@ -43,9 +43,9 @@ func (r *SubscriptionFollowLeagueRepository) GetAllPaginate(pagination *entity.P
 	return pagination, nil
 }
 
-func (r *SubscriptionFollowLeagueRepository) Get(subId, leagueId int) (*entity.SubscriptionFollowLeague, error) {
+func (r *SubscriptionFollowLeagueRepository) Get(subId, leagueId int64) (*entity.SubscriptionFollowLeague, error) {
 	var c entity.SubscriptionFollowLeague
-	err := r.db.Where(&entity.SubscriptionFollowLeague{SubscriptionID: int64(subId), LeagueID: int64(leagueId)}).Take(&c).Error
+	err := r.db.Where(&entity.SubscriptionFollowLeague{SubscriptionID: subId, LeagueID: leagueId}).Take(&c).Error
 	if err != nil {
 		return nil, err
 	}
